refactor(storage): move supabase utils to file service
