+++ backend/src/api/file/file.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { FileService } from './file.service';
+
+@Module({
+  providers: [FileService],
+  exports: [FileService],
+})
+export class FileModule {}
